requirements.txt
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
torch>=2.0
torchvision>=0.15
torchgeo>=0.5
transformers>=4.30
open_clip_torch>=2.20
faiss-cpu>=1.7
tqdm
